# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exit Coach API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Unique scenario identifier
 */
export type ScenarioId = (typeof ScenarioId)[keyof typeof ScenarioId];

export const ScenarioId = {
  performance_issue: "performance_issue",
  layoff: "layoff",
  misconduct: "misconduct",
  pip_failure: "pip_failure",
} as const;

export interface Scenario {
  /** Unique scenario identifier */
  id: ScenarioId;
  name: string;
  description: string;
}

/**
 * Unique persona identifier
 */
export type PersonaId = (typeof PersonaId)[keyof typeof PersonaId];

export const PersonaId = {
  tearful: "tearful",
  defensive: "defensive",
  withdrawn: "withdrawn",
  professional: "professional",
  angry: "angry",
} as const;

export interface Persona {
  /** Unique persona identifier */
  id: PersonaId;
  name: string;
  emotionalStyle: string;
  description: string;
}

export type TurnRole = (typeof TurnRole)[keyof typeof TurnRole];

export const TurnRole = {
  employee: "employee",
  manager: "manager",
} as const;

/**
 * A single conversation turn in the practice session
 */
export interface Turn {
  /**
   * @minimum 1
   * @maximum 5
   */
  turn_index: number;
  role: TurnRole;
  /** @maxLength 2000 */
  transcript: string;
  /** Present on manager turns only */
  coaching_tip?: string;
  /**
   * Employee emotional intensity (1=calm, 10=highly distressed/angry). Present on employee turns only.
   * @minimum 1
   * @maximum 10
   */
  emotion_score?: number;
}

/**
 * Sanitised session state — voice_id is never included
 */
export interface SessionState {
  consent_given: boolean;
  voice_cloned: boolean;
  /** Whether a voice_id exists in the session (the ID itself is never returned) */
  voice_id_present: boolean;
  scenario?: string | null;
  persona?: string | null;
  turns: Turn[];
}

export interface ConsentRequest {
  /** Must be true; server rejects false */
  consentGiven: boolean;
}

export interface ConsentResponse {
  /** Server-side ISO 8601 timestamp of when consent was recorded */
  timestamp: string;
}

export interface CloneVoiceResponse {
  success: boolean;
  /** True when voice cloning failed and a generic ElevenLabs voice will be used instead */
  fallback: boolean;
}

export interface CoachingTipRequest {
  /** Manager's recorded audio turn (WebM, MP4, or OGG) */
  audio: Blob;
  /**
   * Turn number (1–5)
   * @minimum 1
   * @maximum 5
   */
  turnIndex: number;
}

export interface CoachingTipResponse {
  /** Whisper-transcribed text of the manager's turn */
  transcript: string;
  /** GPT-4o-mini coaching tip (strength + improvement) for this turn */
  coachingTip: string;
  /**
   * Simulated employee emotional intensity score
   * @minimum 1
   * @maximum 10
   */
  emotionScore: number;
}

/**
 * A rewritten manager turn with optional audio URL
 */
export interface ImprovedTurn {
  turnIndex: number;
  originalTranscript: string;
  improvedTranscript: string;
  /** URL to the generated TTS audio file (available once generated) */
  audioUrl?: string;
}

export interface FeedbackSummary {
  /** Things the manager did well across the session */
  strengths: string[];
  /** Specific, actionable areas for improvement */
  improvements: string[];
  /** Qualitative overall assessment of the session */
  summary: string;
  /** Emotion scores (1-10) for each employee turn in sequence, used for the Emotion Arc Chart */
  emotionArc: number[];
}

export type UnauthorizedResponse = {
  error: string;
};

export type Ping200 = {
  ok: boolean;
};

export type UpdateSessionBody = {
  scenario?: string;
  persona?: string;
};

export type CloneVoiceBody = {
  /** Manager's voice recording (30-60 seconds of natural speech) */
  audio: Blob;
};