# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Exit Coach API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coaching-tip:
    post:
      operationId: getCoachingTip
      tags: [coaching]
      summary: Transcribe audio and return a coaching tip
      description: |
        Transcribes the manager's recorded turn with OpenAI Whisper,
        then uses GPT-4o-mini to return a coaching tip (strengths + improvement)
        and an emotion score (1-10) for the employee's simulated emotional state.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CoachingTipRequest"
      responses:
        "200":
          description: Transcript, coaching tip, and emotion score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingTipResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /improved-replay:
    post:
      operationId: generateImprovedReplay
      tags: [coaching]
      summary: Generate improved replay of the session
      description: |
        Rewrites all manager turns using GPT-4o-mini, then generates TTS audio
        for each rewritten turn in the manager's cloned voice.
        Progressive generation begins after turn 1 to hide end-of-session latency.
      responses:
        "200":
          description: Array of improved turns with transcripts and audio URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImprovedTurn"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /feedback-summary:
    post:
      operationId: generateFeedbackSummary
      tags: [coaching]
      summary: Generate session feedback summary
      description: |
        Produces a strengths-first coaching summary for the completed session,
        including per-turn recap, areas for improvement, qualitative assessment,
        and an emotion arc array for charting.
      responses:
        "200":
          description: Feedback summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  responses:
    Unauthorized:
      description: No active session
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            required:
              - error
  schemas:
...
    CoachingTipRequest:
      type: object
      properties:
        audio:
          type: string
          format: binary
          description: Manager's recorded audio turn (WebM, MP4, or OGG)
        turnIndex:
          type: integer
          minimum: 1
          maximum: 5
          description: Turn number (1–5)
      required:
        - audio
        - turnIndex
    CoachingTipResponse:
      type: object
      properties:
        transcript:
          type: string
          description: Whisper-transcribed text of the manager's turn
        coachingTip:
          type: string
          description: GPT-4o-mini coaching tip (strength + improvement) for this turn
        emotionScore:
          type: integer
          minimum: 1
          maximum: 10
          description: Simulated employee emotional intensity score
      required:
        - transcript
        - coachingTip
        - emotionScore
    ImprovedTurn:
      type: object
      description: A rewritten manager turn with optional audio URL
      properties:
        turnIndex:
          type: integer
        originalTranscript:
          type: string
        improvedTranscript:
          type: string
        audioUrl:
          type: string
          description: URL to the generated TTS audio file (available once generated)
      required:
        - turnIndex
        - originalTranscript
        - improvedTranscript
    FeedbackSummary:
      type: object
      properties:
        strengths:
          type: array
          items:
            type: string
          description: Things the manager did well across the session
        improvements:
          type: array
          items:
            type: string
          description: Specific, actionable areas for improvement
        summary:
          type: string
          description: Qualitative overall assessment of the session
        emotionArc:
          type: array
          items:
            type: integer
          description: Emotion scores (1-10) for each employee turn in sequence, used for the Emotion Arc Chart
      required:
        - strengths
        - improvements
        - summary
        - emotionArc