# lib/api-zod/src/generated/types/coachingTipRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exit Coach API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CoachingTipRequest {
  /** Manager's recorded audio turn (WebM, MP4, or OGG) */
  audio: Blob;
  /**
   * Turn number (1–5)
   * @minimum 1
   * @maximum 5
   */
  turnIndex: number;
}